requirements.txt
mlflow>=2.0
pyyaml
cairosvg
torch
torchvision
pytorch-lightning
opencv-python
numpy
scipy
scikit-learn
pillow
torchinfo